\name{fib}
\alias{fib}
\title{Fibonnacci Function}
\description{
The classic rubbish fibonnaci function
}
\usage{fib(n)}
\arguments{
\item{n}{the function argument}}

\details{
Calculates fibonnaci numbers using tree-recursion,
making it a very bad function to use ever.
}

\references{just about any dumb-ass compsci book}

\examples{
fib(5)
fib(10)
}

\keyword{math}
